+++ SoftwareRenderer/mymath.cpp
@@ -1,7 +1,7 @@
 #include "mymath.h"
 
-double MyMath::DegreesToRadians(double d)
+float MyMath::DegreesToRadians(float d)
 {
-	return d * (M_PI / 180.0);
+	return d * (static_cast<float>(M_PI) / 180.0f);
 }
